+++ packages/markdown-emitter/src/md/file-generators/module.ts
@@ -2,9 +2,10 @@ import { LinkedFormattedSourceFile, LinkedFormattedSymbol } from '@code-to-json/
 import { isDefined } from '@code-to-json/utils';
 import { Dict } from '@mike-north/types';
 import { SortedExportSymbols, sortSymbols } from '@snap-doc/core';
-import { heading, inlineCode, rootWithTitle, text } from 'mdast-builder';
+import { heading, inlineCode, rootWithTitle, text, paragraph, link } from 'mdast-builder';
 import { Node } from 'unist';
 import { FileEmitterWorkspace, EmitterState } from '@snap-doc/emitter';
+import { resolveAlias } from '@snap-doc/utils';
 import { createDocumentationForCommentData } from '../utils/comment-data';
 import { mdForSymbol } from '../utils/symbol';
 import { addToc } from '../utils/toc';
@@ -25,8 +26,8 @@ function createExportSection(
   state: EmitterState,
   w: FileEmitterWorkspace,
   sectionName: string,
-  filePath: string,
   symCollection: Dict<LinkedFormattedSymbol>,
+  file: LinkedFormattedSourceFile,
   { summarizeOnly }: ExportSectionOptions,
 ): Node[] {
   const symbols = Object.keys(symCollection)
@@ -40,14 +41,43 @@ function createExportSection(
       if (!sym) {
         return all;
       }
-      all.push(
-        ...mdForSymbol(state, w, sym, {
-          includeTitle: true,
-          includeDetails: !summarizeOnly,
-          baseDepth: 4,
-          path: filePath,
-        }),
-      );
+
+      const resolvedSym = resolveAlias(sym);
+      const { sourceFile } = resolvedSym;
+      if (!sourceFile || sourceFile.id === file.id) {
+        all.push(
+          ...mdForSymbol(state, w, sym, file, {
+            includeTitle: true,
+            includeDetails: !summarizeOnly,
+            baseDepth: 4,
+          }),
+        );
+      } else {
+        all.push(
+          heading(
+            4,
+            link(
+              w.relativePath(state, file, resolvedSym, 'md'),
+              resolvedSym.text || resolvedSym.name,
+              inlineCode(resolvedSym.text || resolvedSym.name),
+            ),
+          ),
+          paragraph([
+            text('re-export of symbol '),
+            link(
+              w.relativePath(state, file, resolvedSym, 'md'),
+              resolvedSym.name,
+              inlineCode(resolvedSym.name),
+            ),
+            text(' from '),
+            link(
+              w.relativePath(state, file, sourceFile, 'md'),
+              sourceFile.moduleName,
+              inlineCode(sourceFile.moduleName),
+            ),
+          ]),
+        );
+      }
       return all;
     },
     [] as Node[],
@@ -61,21 +91,17 @@ export function createExportSections(
   state: EmitterState,
   w: FileEmitterWorkspace,
   { classes, properties, types, functions }: SortedExportSymbols,
-  filePath: string,
+  file: LinkedFormattedSourceFile,
   options: MarkdownGenOptions,
 ): Node[] {
   const parts: Node[] = [
-    ...createExportSection(state, w, 'Properties', filePath, properties, { summarizeOnly: false }),
-    ...createExportSection(state, w, 'Functions', filePath, functions, { summarizeOnly: false }),
-    ...createExportSection(state, w, 'Types', filePath, types, {
+    ...createExportSection(state, w, 'Properties', properties, file, { summarizeOnly: false }),
+    ...createExportSection(state, w, 'Functions', functions, file, { summarizeOnly: false }),
+    ...createExportSection(state, w, 'Types', types, file, {
       summarizeOnly: !options.detailedModules,
-      headerUrlFactory: (sym: LinkedFormattedSymbol) =>
-        w.host.pathRelativeTo(w.host.combinePaths(filePath, '..'), w.pathFor(state, sym)),
     }),
-    ...createExportSection(state, w, 'Classes', filePath, classes, {
+    ...createExportSection(state, w, 'Classes', classes, file, {
       summarizeOnly: !options.detailedModules,
-      headerUrlFactory: (sym: LinkedFormattedSymbol) =>
-        w.host.pathRelativeTo(w.host.combinePaths(filePath, '..'), w.pathFor(state, sym)),
     }),
   ];
   const secRoot = [heading(2, text('Exports')), ...parts];
@@ -104,10 +130,7 @@ export function markdownForSourceFile(
   rootNode.children.push(...createDocumentationForCommentData(documentation));
   if (exportSymbols && Object.keys(exportSymbols).length > 0) {
     const sortedExports: SortedExportSymbols = sortSymbols(exportSymbols);
-    const filePath = workspace.pathFor(state, file);
-    rootNode.children.push(
-      ...createExportSections(state, workspace, sortedExports, filePath, options),
-    );
+    rootNode.children.push(...createExportSections(state, workspace, sortedExports, file, options));
     symbolsToSerialize.classes.push(
       ...Object.keys(sortedExports.classes)
         .map(k => sortedExports.classes[k])