+++ packages/markdown-emitter/src/md/utils/symbol.ts
@@ -11,19 +11,20 @@ import {
   heading,
   inlineCode,
   link,
-  paragraph,
-  text,
   list,
   listItem,
+  paragraph,
+  separator,
   strong,
+  text,
 } from 'mdast-builder';
 import { Node } from 'unist';
 import { FileEmitterWorkspace, EmitterState } from '@snap-doc/emitter';
+import { Pathable } from '@snap-doc/emitter';
 import { createDocumentationForCommentData } from './comment-data';
 import { mdSignatures } from './signature';
 
 export interface MDSymbolOptions {
-  path: string;
   includeDetails?: boolean;
   includeTitle?: boolean;
   baseDepth?: number;
@@ -60,7 +61,7 @@ function mdForSymbolSignatures(
       parts.push(...head);
     }
 
-    parts.push(...mdSignatures(s, sigs));
+    parts.push(strong(text('Signatures')), ...mdSignatures(s, sigs));
   }
   return parts;
 }
@@ -72,7 +73,7 @@ function mdForProperties(props: Dict<LinkedFormattedSymbol>, opts: MDSymbolOptio
     .reduce(
       (lst, p) => {
         lst.push(
-          heading((opts.baseDepth || 1) + 1, text(p.name)),
+          heading((opts.baseDepth || 1) + 1, text(p.text || p.name)),
           blockquote(() => {
             const parts: Node[] = [];
             const { valueType, documentation } = p;
@@ -170,9 +171,10 @@ export function mdForSymbolDetails(s: LinkedFormattedSymbol, opts: MDSymbolOptio
 }
 
 function mdForBaseTypes(
-  _w: FileEmitterWorkspace,
-  _path: string,
+  state: EmitterState,
+  w: FileEmitterWorkspace,
   _s: LinkedFormattedSymbol,
+  pageItem: Pathable,
   type: LinkedFormattedType | undefined,
 ): Node {
   if (!type || !type.baseTypes) {
@@ -183,20 +185,25 @@ function mdForBaseTypes(
     text(' '),
     inlineCode('extends'),
     text(' '),
-    ...type.baseTypes.reverse().map(
-      bt => (bt.symbol ? inlineCode(bt.symbol.name) : inlineCode('(unknown symbol)')),
-      // bt.symbol
-      // ? link(
-      //     w.host.pathRelativeTo(w.host.combinePaths(w.pathFor(s)), path),
-      //     bt.symbol.name,
-      //     inlineCode(bt.symbol.name)
-      //   )
-      // : inlineCode('(unknown symbol)')
+    ...type.baseTypes.reverse().map(bt =>
+      // bt.symbol ? inlineCode(bt.symbol.text || bt.symbol.name) : inlineCode('(unknown symbol)'),
+      bt.symbol
+        ? link(
+            w.relativePath(state, pageItem, bt.symbol, 'md'),
+            bt.symbol.text || bt.symbol.name,
+            inlineCode(bt.symbol.text || bt.symbol.name),
+          )
+        : inlineCode('(unknown symbol)'),
     ),
   ]);
 }
 
-function mdForSymbolHeader(w: FileEmitterWorkspace, path: string, s: LinkedFormattedSymbol): Node {
+function mdForSymbolHeader(
+  state: EmitterState,
+  w: FileEmitterWorkspace,
+  s: LinkedFormattedSymbol,
+  pageItem: Pathable,
+): Node {
   const { flags, type, accessModifier } = s;
   const kids: Node[] = [];
   if (accessModifier) {
@@ -216,7 +223,7 @@ function mdForSymbolHeader(w: FileEmitterWorkspace, path: string, s: LinkedForma
       ),
     );
   }
-  kids.push(mdForBaseTypes(w, path, s, type));
+  kids.push(mdForBaseTypes(state, w, s, pageItem, type));
 
   if (kids.length === 0) {
     return text('');
@@ -228,13 +235,14 @@ export function mdForSymbol(
   state: EmitterState,
   w: FileEmitterWorkspace,
   s: LinkedFormattedSymbol,
+  pageItem: Pathable,
   opts: MDSymbolOptions,
 ): Node[] {
   const { documentation } = s;
 
-  const parts: Node[] = [text('---')];
+  const parts: Node[] = [separator];
   parts.push(...mdForSymbolTitle(state, w, s, opts));
-  parts.push(mdForSymbolHeader(w, opts.path, s));
+  parts.push(mdForSymbolHeader(state, w, s, pageItem));
   if (documentation) {
     parts.push(paragraph(createDocumentationForCommentData(documentation)));
   }