+++ packages/types/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.14](https://github.com/snap-doc/snap-doc/compare/@snap-doc/types@0.5.13...@snap-doc/types@0.5.14) (2019-03-09)
+
+**Note:** Version bump only for package @snap-doc/types
+
+
+
+
+
+## [0.5.13](https://github.com/snap-doc/snap-doc/compare/@snap-doc/types@0.5.12...@snap-doc/types@0.5.13) (2019-03-07)
+
+**Note:** Version bump only for package @snap-doc/types
+
+
+
+
+
+## [0.5.12](https://github.com/snap-doc/snap-doc/compare/@snap-doc/types@0.5.11...@snap-doc/types@0.5.12) (2019-03-06)
+
+**Note:** Version bump only for package @snap-doc/types
+
+
+
+
+
+## [0.5.11](https://github.com/snap-doc/snap-doc/compare/@snap-doc/types@0.5.10...@snap-doc/types@0.5.11) (2019-03-02)
+
+**Note:** Version bump only for package @snap-doc/types
+
+
+
+
+
+## [0.5.10](https://github.com/snap-doc/snap-doc/compare/@snap-doc/types@0.5.9...@snap-doc/types@0.5.10) (2019-03-01)
+
+**Note:** Version bump only for package @snap-doc/types
+
+
+
+
+
+## [0.5.9](https://github.com/snap-doc/snap-doc/compare/@snap-doc/types@0.5.8...@snap-doc/types@0.5.9) (2019-02-28)
+
+**Note:** Version bump only for package @snap-doc/types
+
+
+
+
+
 ## [0.5.8](https://github.com/snap-doc/snap-doc/compare/@snap-doc/types@0.5.7...@snap-doc/types@0.5.8) (2019-02-24)
 
 **Note:** Version bump only for package @snap-doc/types